requirements2.txt
coc.py
ruamel.yaml